requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.4.1
Flask-WTF==1.2.2
Werkzeug==2.0.2
firebase-admin==6.6.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
python-dotenv==1.0.1
Pillow==11.0.0
pandas==2.2.3
numpy==2.2.0
scikit-learn==1.6.0
thefuzz==0.22.1